Added access control

+++ routes/articles.js
@@ -1,14 +1,27 @@
 const express = require('express');
 const router = express.Router();
 
-//bring in Article Model
+//Article Model
 let Article = require('../models/article');
+//user model
+let User = require('../models/user');
+
+
+//Access control
+const ensureAuthenticated = (req, res, next) => {
+    if(req.isAuthenticated()) {
+        return next();
+    } else {
+        req.flash('danger', "Please login");
+        res.redirect('/users/login');
+    }
+}
 
 
 // Add submit POST route
 router.post('/add', (req, res) => {
     req.checkBody('title', 'Title is required').notEmpty();
-    req.checkBody('author', 'author is required').notEmpty();
+    // req.checkBody('author', 'author is required').notEmpty();
     req.checkBody('body', 'body is required').notEmpty();
 
     let errors = req.validationErrors();
@@ -21,7 +34,7 @@ router.post('/add', (req, res) => {
     } else {
         let article = new Article();
         article.title = req.body.title;
-        article.author = req.body.author,
+        article.author = req.user._id,
         article.body = req.body.body
     
         article.save((err) => {
@@ -38,7 +51,7 @@ router.post('/add', (req, res) => {
 })
 
 //Add route
-router.get('/add', (req, res) => {
+router.get('/add', ensureAuthenticated, (req, res) => {
     res.render('add_article', {
         title: 'Add Articles'
     })
@@ -46,8 +59,12 @@ router.get('/add', (req, res) => {
 
 
 //load edit form
-router.get('/edit/:id', (req, res) => {
+router.get('/edit/:id', ensureAuthenticated, (req, res) => {
     Article.findById(req.params.id, (err, article) => {
+        if(article.author != req.user._id) {
+            req.flash('danger', 'Not Authorized');
+            return res.redirect('/');
+        }
         res.render('edit_article', {
             title: 'Edit Article',
             article: article
@@ -77,22 +94,36 @@ router.post('/edit/:id', (req, res) => {
 
 
 router.delete('/:id', (req, res) => {
+    if(!req.user._id) {
+        res.status(500).send();
+    }
     let query = {_id: req.params.id}
 
-    Article.deleteOne(query, (err) => {
-        if(err) {
-            console.log(err);
+    Article.findById(req.params.id, function(err, article) {
+        if(article.author != req.user._id) {
+            res.status(500).send();
             return;
+        } else {
+            Article.deleteOne(query, (err) => {
+                if(err) {
+                    console.log(err);
+                    return;
+                }
+                res.send('Success');
+            });
         }
-        res.send('Success');
-    })
+    });
+
 });
 
 // get single article
 router.get('/:id', (req, res) => {
     Article.findById(req.params.id, (err, article) => {
-        res.render('article', {
-            article: article
+        User.findById(article.author, (err, user) => {
+            res.render('article', {
+                article: article,
+                author: user.name
+            });
         });
     });
 });
